+++ include/hip.h
@@ -95,7 +95,7 @@ class Hip
             Atomic::clr_mask<true>(hip()->api_flg, static_cast<typeof hip()->api_flg>(f));
         }
 
-        static bool cpu_online (unsigned cpu)
+        static bool cpu_online (unsigned long cpu)
         {
             return cpu < NUM_CPU && hip()->cpu_desc[cpu].flags & 1;
         }
